lib wrapper re-write
