+++ module.json
@@ -2,7 +2,7 @@
     "name": "Next-Up",
     "title": "Next Up",
     "description": "Several Combat Enhancements, pan to combatant, open sheets and turn markers.",
-    "version": "0.0.56",
+    "version": "0.0.57",
     "author" : "Kandashi",
     "manifestPlusVersion": "1.0.0",
     "authors": [
@@ -14,7 +14,7 @@
       }
     ],
     "esmodules": [
-      "/src/combatFocus.js"        
+      "/src/nextUp.js"        
   ],
   "scripts": [
     "greensock/dist/gsap.min.js"
